Overhaul tests to use a common ExceedGutTest parent
* Apply some class_names

* Proof of concept for pulling all common test functionality into a superclass

* Refactor tests for Yuzuriha

* Waldstein

* Remake core exceed test features

Test card IDs are now generated automatically with an incrementing counter, instead of being
limited to the allotment hard-coded into the test suite. The downside is that you have to
assign the ids to variables if you ever want to refer to them again. The upside is
this guarantees no collisions on reused card IDs.

Player choices during a strike are now all listed in a single array per player. It is up to
the test author to ensure that the arguments provided are appropriate to the sequence
of decisions that will be presented during the test case. The upshot is that whenever a new
kind of decision needs to be made for a test, its parameters can simply be added to the
decisions array, and handling for that decision type is easy to add through the match
statements in process_decisions and process_remaining_decisions. This replaces the need to
alter the function signature of execute_strike with new arguments each time a new kind of
decision needs to be supported.

* Documentation

* Change test card ID assignment paradigm

* Convert Yuzu

* Convert Waldstein (checkpoint)

* Uncomment Waldstein

* Waldstein checkpoint

* Finish Waldstein

* Revert enums.gd

* Update some comments